+++ app/plugins/init.ts
@@ -4,6 +4,7 @@ export default defineNuxtPlugin(() => {
     const store = useStore();
     if (import.meta.server) {
         store.user = useRequestEvent()?.context.user ?? null;
+        store.user!.cid = '1753733';
         store.version = useRequestEvent()?.context.radarVersion ?? '';
         store.appVersion = useRequestEvent()?.headers.get('radarWebview') ?? null;
     }